Changed underlying operand type to double long for more precision. Removed dead code.

+++ Calculator/CalcOperation.h
@@ -14,7 +14,7 @@
 
 using std::string;
 
-using OperandDataType = float;
+using OperandDataType = long double;
 
 class CalcOperation
 {
